pyproject.toml
[tool.poetry]
name = "pokemon-tcg-bot"
version = "1.0.0"
description = ""
authors = []
package-mode = false

[tool.poetry.dependencies]
python = ">3.11,<3.12"
bs4 = "0.0.2"
pandas = "2.2.2"
numpy = "1.25.2"
rembg = "2.0.59"
pillow = "10.4.0"
matplotlib = "3.9.2"
discord-py = "2.4.0"
reactionmenu = "3.1.7"
python-dotenv = "1.0.1"
ImageHash = "4.3.1"
SQLAlchemy = "2.0.35"
psycopg2-binary = "2.9.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"